requirements.txt
boto3==1.35.87
Flask==3.1.0
gunicorn==23.0.0
pathlib==1.0.1
Pillow==9.5.0
pytesseract==0.3.13
requests==2.32.3
scikit-learn==1.4.1.post1
sympy==1.13.1
tflearn @ git+https://github.com/MihaMarkic/tflearn.git@6472b8588e758ff4a33a2764d4ee638bbd0e42f0
torch==2.5.1
Werkzeug==3.1.3
transformers==4.27.1
datasets==2.10.1
streamlit==1.14.0
